Built-in basic functionality

+++ Mydemenageur.API/Entities/Mover.cs
@@ -13,7 +13,6 @@ namespace Mydemenageur.API.Entities
         [BsonId]
         [BsonRepresentation(BsonType.ObjectId)]
         public string Id { get; set; }
-
         /// <summary>
         /// The id of the user associated with the mover
         /// </summary>
@@ -28,12 +27,11 @@ namespace Mydemenageur.API.Entities
         /// Average customer of the mover's
         /// </summary>
         /// <example>8.5</example>
-        public float AverageCustomerRating { get; set; }
+        public double AverageCustomerRating { get; set; }
         [BsonRepresentation(BsonType.ObjectId)]
         /// <summary>
         /// The file is list
         /// </summary>
         public List<string> FileIds { get; set; }
-
     }
 }
